+++ packages/leancode_force_update/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.1-alpha
+- Fix: don't check for updates when the app is not in the foreground
+
 ## 1.0.0-alpha
 
 Initial release